pyproject.toml
[project]
name = "testing-repository-simple-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask==2.2.3",
    "flask-cors==4.0.0",
    "flask-jwt-extended==4.4.4",
    "gevent>=24.11.1",
    "psutil>=6.1.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = ["black>=25.1.0", "flake8>=7.1.1", "reorder-python-imports>=3.14.0"]
